+++ client/components/interview-board/interview-board.css
@@ -2,7 +2,7 @@
     box-sizing: border-box;
    }
 
-   html, body, #app, #rootDiv {
+   html, body, #app, #rootDiv, div {
     height: 100%;
 }
 
@@ -41,6 +41,7 @@
     padding: 0;
     width: 100%;
 }
+
 .texty {
 width: 100%;
 box-sizing: border-box;
@@ -51,78 +52,10 @@ background-color: black;
 color:white;
 padding: 1em;
 }
+
 .mycanvas {
-width: 1400px;
-height:1000px;
+width: 2000px;
+height: 2000px;
 background-color: white;
 }
 
-   /*=====================*/
-.canvas {
-    height: 50%;
-}
-
-.editor {
-    height: 50%;
-}
-
-.textArea {
-    resize: vertical;
-    height: 50%;
-    width: 100%;
-    min-height: 10%;
-    max-height: 90%;
-    background-color: black;
-    color:white;   
-    position: absolute;
-    outline: none;
-}
-
-.outer {
-    height: 400px;
-}
-
-.Resizer {
-    background: #000;
-    opacity: .2;
-    z-index: 1;
-    -moz-box-sizing: border-box;
-    -webkit-box-sizing: border-box;
-    box-sizing: border-box;
-    -moz-background-clip: padding;
-    -webkit-background-clip: padding;
-    background-clip: padding-box;
-}
-
- .Resizer:hover {
-    -webkit-transition: all 2s ease;
-    transition: all 2s ease;
-}
-
- .Resizer.horizontal {
-    height: 11px;
-    margin: -5px 0;
-    border-top: 5px solid rgba(255, 255, 255, 0);
-    border-bottom: 5px solid rgba(255, 255, 255, 0);
-    cursor: row-resize;
-    width: 100%;
-}
-
-.Resizer.horizontal:hover {
-    border-top: 5px solid rgba(0, 0, 0, 0.5);
-    border-bottom: 5px solid rgba(0, 0, 0, 0.5);
-}
-
-.Resizer.vertical {
-    width: 11px;
-    margin: 0 -5px;
-    border-left: 5px solid rgba(255, 255, 255, 0);
-    border-right: 5px solid rgba(255, 255, 255, 0);
-    cursor: col-resize;
-    height: 100%;
-}
-
-.Resizer.vertical:hover {
-    border-left: 5px solid rgba(0, 0, 0, 0.5);
-    border-right: 5px solid rgba(0, 0, 0, 0.5);
-}
