+++ client/components/questions/questions.css
@@ -1,0 +1,3 @@
+.ui.attached.tabular.menu {
+  height: 30px;
+}
