+++ client/store/index.js
@@ -8,8 +8,9 @@ import jobs from './jobs'
 import whiteboard from '../components/interview-board/whiteboard-reducer'
 import textarea from '../components/interview-board/textarea-reducer'
 import questions from '../components/questions/questions-reducer'
+import saved from '../components/interview-container/save-state-reducer'
 
-const reducer = combineReducers({user, whiteboard, textarea, jobs, questions})
+const reducer = combineReducers({user, whiteboard, textarea, jobs, questions, saved})
 
 
 const middleware = composeWithDevTools(applyMiddleware(
